Abstract out new hook: usePaginationState
